+++ src/main/java/nadiendev/voidminers/world/block/entity/SolarControllerBE.java
@@ -103,7 +103,7 @@ public class SolarControllerBE extends BlockEntity {
 
                 MiscUtil.getNeededBlocks(MiscUtil.structureMap.get(structure.toString())).forEach((string, integer) -> {
                     tooltip.add(Component.literal("• ").withStyle(ChatFormatting.GRAY)
-                            .append(Component.literal(string.contains("Null") ? "Modifiers" : string).withStyle(ChatFormatting.WHITE))
+                            .append(Component.literal(string.contains("Null") ? "Modifier" : string).withStyle(ChatFormatting.WHITE))
                             .append(Component.literal(": ").withStyle(ChatFormatting.GRAY))
                             .append(Component.literal(String.valueOf(integer)).withStyle(ChatFormatting.RED)));
                 });